move tar commands to archiver.tar
